moved alloc setup behind `prelude::init`
