+++ board/lib/src/lib.rs
@@ -1,19 +1,10 @@
-#[global_allocator]
-static HEAP: Heap = Heap::empty();
-const HEAP_SIZE: usize = 128;
-static mut HEAP_MEM: [MaybeUninit<u8>; HEAP_SIZE] = [MaybeUninit::uninit(); HEAP_SIZE];
+#![cfg_attr(not(test), no_std)]
+#![feature(alloc_error_handler)] // required for our prelude that sets up the heap
+#![feature(prelude_2024)]
+#![no_main]
 
-// Initialize the allocator BEFORE you use it
-fn alloc_heap() {
-    {
-        unsafe { HEAP.init(HEAP_MEM.as_ptr() as usize, HEAP_SIZE) }
-    }
-}
 
-#[alloc_error_handler]
-fn oom(_: Layout) -> ! {
-    loop {}
-}
+pub mod prelude;
 
 // TODO:: import the libraries needed and update these FFI functions
 // #[no_mangle]
